rename recorder *-new files to overwrite old versions, change imports to remove *-new from filenames
