+++ app/components/exp-lookit-calibration/component.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import layout from './template';
 import ExpFrameBaseComponent from '../exp-frame-base/component';
-import VideoRecord from '../../mixins/video-record-new';
+import VideoRecord from '../../mixins/video-record';
 import PauseUnpause from '../../mixins/pause-unpause';
 import ExpandAssets from '../../mixins/expand-assets';
 import isColor from '../../utils/is-color';
