+++ app/components/exp-lookit-video/component.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import layout from './template';
 import ExpFrameBaseComponent from '../exp-frame-base/component';
-import VideoRecord from '../../mixins/video-record-new';
+import VideoRecord from '../../mixins/video-record';
 import ExpandAssets from '../../mixins/expand-assets';
 import isColor, {colorSpecToRgbaArray, textColorForBackground} from '../../utils/is-color';
 import { audioAssetOptions, videoAssetOptions } from '../../mixins/expand-assets';
