+++ app/mixins/video-record.js
@@ -28,10 +28,20 @@ let {
  * You will also need to set `recorderElement` if the recorder is to be housed other than
  * in an element identified by the ID `recorder`.
  *
- * The properties `recorder`, `videoList`, `stoppedRecording`, `recorderReady`, and
- * `videoId` become available to the consuming frame. The recorder object has fields
- * that give information about its state: `hasWebCam`, 'hasCamAccess`, `recording`,
- * `connected`, and `micChecked` - for details, see services/video-recorder.js. These
+ * The following properties become available to the consuming frame:
+ *  recorder
+ *  videoList
+ *  stoppedRecording
+ *  recorderReady
+ *  videoId
+ * The recorder object has the following fields that give information about its state: 
+ *  hasWebCam
+ *  hasCamAccess
+ *  recording
+ *  connected
+ *  micChecked
+ * For details, see services/video-recorder.js. 
+ * These
  * can be accessed from the consuming frame as e.g. `this.get('recorder').get('hasWebCam')`.
  *
  * If starting recording automatically,  the function `onRecordingStarted` will be called
@@ -42,8 +52,7 @@ let {
  *
  * Events recorded in a frame that extends VideoRecord will automatically have additional
  * fields videoId (video filename), pipeId (temporary filename initially assigned by
- * the recording service),
- * and streamTime (when in the video they happened, in s).
+ * the recording service), and streamTime (when in the video they happened, in s).
  *
  * Setting up the camera is handled in didInsertElement, and making sure recording is
  * stopped is handled in willDestroyElement (Ember hooks that fire during the component
@@ -150,6 +159,14 @@ export default Ember.Mixin.create({
      * @default 0
      */
     audioOnly: 0,
+
+    /**
+     * Whether to do check the mic input during recorder set up (before the install promise is resolved). 
+     * Defaults to false and can be overridden by consuming frame.
+     * @property {Boolean} checkMic
+     * @default false
+     */
+    checkMic: false,
 
     /**
      * Whether to use the camera in this frame. Consuming frame should set this property
@@ -291,16 +308,14 @@ export default Ember.Mixin.create({
         const videoId = this._generateVideoId();
         this.set('videoId', videoId);
         const recorder = new VideoRecorder({element: element});
-        const pipeLoc = Ember.getOwner(this).resolveRegistration('config:environment').pipeLoc;
-        const pipeEnv = Ember.getOwner(this).resolveRegistration('config:environment').pipeEnv;
-        console.log('audio only: ', this.get('audioOnly'));
-        const installPromise = recorder.install(this.get('videoId'), pipeLoc, pipeEnv,
-            this.get('maxRecordingLength'), this.get('autosave'), this.get('audioOnly'));
+        const installPromise = recorder.install(this.get('videoId'), 
+            this.get('maxRecordingLength'), this.get('autosave'), this.get('audioOnly'), this.get('checkMic'));
 
-        // Track specific events for all frames that use  VideoRecorder
+        // Track specific events for all frames that use VideoRecorder
         var _this = this;
         recorder.on('onCamAccess', (recId, hasAccess) => {   // eslint-disable-line no-unused-vars
             console.log('video record mixin: on cam access fired');
+            _this.get('recorder').set('hasCamAccess', hasAccess);
             if (!(_this.get('isDestroyed') || _this.get('isDestroying'))) {
                 _this.send('setTimeEvent', 'recorder.hasCamAccess', {
                     hasCamAccess: hasAccess
@@ -319,6 +334,7 @@ export default Ember.Mixin.create({
         this.send('setTimeEvent', 'setupVideoRecorder', {
             videoId: videoId
         });
+        
         return installPromise;
     },
 
@@ -334,7 +350,7 @@ export default Ember.Mixin.create({
             return recorder.record().then(() => {
                 console.log('video record mixin: record promise fulfilled (start recorder function)');
                 this.send('setTimeEvent', 'startRecording', {
-                    pipeId: recorder.get('pipeVideoName')
+                    pipeId: recorder.get('videoName') // TO DO: change 'pipeId' to something else 
                 });
                 if (this.get('videoList') == null) {
                     this.set('videoList', [this.get('videoId')]);
@@ -526,7 +542,7 @@ export default Ember.Mixin.create({
         console.log('video record mixin: when possible to record observer');
         if (this.get('doUseCamera') && this.get('startRecordingAutomatically')) {
             var _this = this;
-            if (this.get('recorder.hasCamAccess') && this.get('recorderReady')) {
+            if (this.get('recorder.hasCamAccess') && this.get('recorderReady') && !(this.get('recording'))) {
                 this.startRecorder().then(() => {
                     console.log('video record mixin: start recorder fulfilled (when possible to record observer function)');
                     _this.set('recorderReady', false);
